온보딩 알럿 추가 및 안쓰는 파일 삭제
