--- presentation/src/main/java/com/keeply/presentation/ui/scan/ScreenshotScreen.kt
@@ -1,129 +1,0 @@
-package com.keeply.presentation.ui.scan
-
-import android.net.Uri
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.material3.Icon
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.paging.compose.collectAsLazyPagingItems
-import coil.compose.rememberAsyncImagePainter
-import com.keeply.presentation.core.components.ImageFrame
-import com.keeply.presentation.core.components.KeeplyAppBar
-import com.keeply.presentation.core.components.KeeplyText
-import com.keeply.presentation.core.theme.KeeplyTheme
-import com.keeply.presentation.core.theme.neutral100
-
-
-@Composable
-fun ScreenshotScreen(
-    onBack: () -> Unit,
-    onNavigateToDetail: (Uri) -> Unit,
-    viewModel: ScanViewModel = hiltViewModel()
-) {
-    val lazyPagingItems = viewModel.screenshots.collectAsLazyPagingItems()
-
-    // TODO: 화면 재진입 시 스샷목록 새로고침
-
-    ScreenshotFrame(
-        count = lazyPagingItems.itemCount,
-        onBack = onBack
-    ) {
-        LazyVerticalGrid(
-            columns = GridCells.Fixed(3),
-            modifier = Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.spacedBy(5.dp),
-            horizontalArrangement = Arrangement.spacedBy(7.dp),
-            contentPadding = PaddingValues(horizontal = 16.dp)
-        ) {
-            items(lazyPagingItems.itemCount) { index ->
-                val isFirstRow = index < 3
-
-                lazyPagingItems[index]?.let { screenshot ->
-                    ImageFrame(
-                        painter = rememberAsyncImagePainter(screenshot.uri),
-                        modifier = Modifier
-                            .padding(top = if (isFirstRow) 28.dp else 0.dp)
-                            .clickable { onNavigateToDetail(screenshot.uri) }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun ScreenshotFrame(
-    modifier: Modifier = Modifier,
-    onBack: () -> Unit,
-    count: Int = 0,
-    content: @Composable ColumnScope.() -> Unit
-) {
-    Column(
-        modifier = modifier
-            .fillMaxHeight()
-            .background(neutral100)
-    ) {
-        KeeplyAppBar(
-            customTitleContent = {
-                Row(
-                    modifier = Modifier
-                        .padding(start = 12.dp)
-                        .align(Alignment.CenterStart),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalAlignment = Alignment.Bottom
-                ) {
-                    KeeplyText(
-                        text = "Screenshots",
-                        style = KeeplyTheme.typography.header03,
-                    )
-
-                    KeeplyText(
-                        text = "$count",
-                        style = KeeplyTheme.typography.header04,
-                        color = KeeplyTheme.colors.neutral500
-                    )
-                }
-            },
-            leadingIcon = {
-                Icon(
-                    painter = KeeplyTheme.icons.chevronLeft,
-                    contentDescription = "Back",
-                    tint = KeeplyTheme.colors.neutralBlack
-                )
-            },
-            onClickLeading = onBack
-        )
-
-        content()
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun ScanScreenPreview() {
-    KeeplyTheme {
-        ScreenshotFrame(
-            onBack = { }
-        ) {
-            KeeplyText(
-                text = "Contents",
-                style = KeeplyTheme.typography.header03,
-            )
-        }
-    }
-}
