+++ presentation/src/main/java/com/keeply/presentation/ui/scan/scanbefore/ScanBeforeViewModel.kt
@@ -3,9 +3,15 @@ package com.keeply.presentation.ui.scan.scanbefore
 import android.net.Uri
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
+import com.keeply.domain.usecase.scan.GetScanOnBoardingVisibilityUseCase
+import com.keeply.domain.usecase.scan.SetDoNotShowDialogUseCase
 import com.keeply.presentation.ui.scan.navigation.ScanRoute
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.launch
 import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.viewmodel.container
@@ -13,7 +19,9 @@ import javax.inject.Inject
 
 @HiltViewModel
 class ScanBeforeViewModel @Inject constructor(
-    private val savedStateHandle: SavedStateHandle
+    private val savedStateHandle: SavedStateHandle,
+    getScanOnBoardingVisibilityUseCase: GetScanOnBoardingVisibilityUseCase,
+    private val setDoNotShowDialogUseCase: SetDoNotShowDialogUseCase
 ):ContainerHost<ScanBeforeState, ScanBeforeSideEffect>, ViewModel() {
     val scanBefore: ScanRoute.ScanBefore = savedStateHandle.toRoute()
 
@@ -22,4 +30,13 @@ class ScanBeforeViewModel @Inject constructor(
             Uri.decode(scanBefore.url)
         ))
 
+    // 온보딩 모달 노출 여부
+    val showOnBoardingModal = getScanOnBoardingVisibilityUseCase()
+        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), true)
+
+    fun onDoNotShowAgain() {
+        viewModelScope.launch {
+            setDoNotShowDialogUseCase(true)
+        }
+    }
 }
